committed temporary - INCOMPLETE: prepared additional module events "OnPlayerConnect", "OnPlayerDisconnect", "OnChat" - INCOMPLETE: added lib/rconcommand.py module - added command.json setting - provided an option to cancel a scheduled restart (for instance through chat command '!cancel' - check commands.json) - new config option "commands" to add a setting file containing possible chat commands and admins - added optional "pyrcon.py" command argument "--gui" to provide a graphical interface (later)

+++ pyrcon.py
@@ -13,8 +13,10 @@ pid = str(os.getpid())
 
 parser = argparse.ArgumentParser(description='Python Rcon cmdlet for ARMA3 Servers')
 parser.add_argument('configfile', help='configuration file in JSON')
-
+parser.add_argument('-g','--gui', action='store_true',help='open the GUI - no other events are enabled')
 args = parser.parse_args()
+
+GUI = args.gui
 
 if not os.path.isfile(args.configfile):
     print ''
@@ -27,13 +29,7 @@ with open(args.configfile) as json_config:
 _bstr = os.path.basename(args.configfile)
 _n = 48 - len(_bstr)
 
-print 'PyRcon for ARMA3 CLI v0.1'
-print ''
-print '#'.ljust(71, '#')
-print '# Configuration file: {}{}#'.format(_bstr , ''.ljust(_n, ' '))
-print '#'.ljust(71, '#')
-print '# DO NOT FORGET TO SETUP THE beserver.cfg LOCATED IN battleye/ FOLDER #'
-print '#'.ljust(71, '#')
+print 'PyRcon for ARMA3 CLI v0.1 using configuration {}'.format(_bstr)
 print ''
 
 # Logging tool configuration
@@ -41,17 +37,16 @@ print 'Logging to: {}'.format(config['logfile'])
 FORMAT = '%(asctime)-15s %(levelname)s %(message)s'
 logging.basicConfig(filename=config['logfile'], level=config['loglevel'], format=FORMAT)
 
-pidfile = '/tmp/pyrcon.{}.run'.format(config['server']['port'])
+if not(GUI):
+    pidfile = '/tmp/pyrcon.{}.run'.format(config['server']['port'])
 
-if os.path.isfile(pidfile):
-    _tmp = 'pyrcon is already running for {}:{}'.format(config['server']['host'], config['server']['port'])
-    print _tmp
-    logging.info(_tmp)
-    exit()
+    if os.path.isfile(pidfile):
+	_tmp = 'pyrcon is already running for {}:{}'.format(config['server']['host'], config['server']['port'])
+	print _tmp
+	logging.info(_tmp)
+	exit()
 
-open(pidfile, 'w').write(pid)
-
-logging.debug("Initialize rconprotocol class object")
+    open(pidfile, 'w').write(pid)
 
 ##
 # Initialize the rconprotocol class
@@ -61,21 +56,32 @@ rcon = Rcon(config['server']['host'], config['server']['rcon_password'], config[
 ##
 # Load the rconrestart module and setup from configuration
 ##
-modRestart = rcon.loadmodule('rconrestart', 'RconRestart', ['RestartMessage'])
-_rlist = []
-for m in config['restart']['messages']:
-    _rlist.append( lib.rconrestart.RestartMessage(m[0],m[1]) )
+if not(GUI):
+    modRestart = rcon.loadmodule('rconrestart', 'RconRestart', ['RestartMessage'])
+    _rlist = []
+    for m in config['restart']['messages']:
+        _rlist.append( lib.rconrestart.RestartMessage(m[0],m[1]) )
 
-modRestart.setMessages( _rlist )
-modRestart.setInterval( config['restart']['interval'] )
-modRestart.setExitOnRestart(config['restart']['exitonrestart'])
+    modRestart.setMessages( _rlist )
+    modRestart.setInterval( config['restart']['interval'] )
+    modRestart.setExitOnRestart(config['restart']['exitonrestart'])
 
 ##
 # Load the rconmessage module and setup from configuration
 ##
-modMessage = rcon.loadmodule('rconmessage', 'RconMessage')
-modMessage.setInterval( config['repeatMessage']['interval'] )
-modMessage.setMessages( config['repeatMessage']['messages'] )
+if not(GUI):
+    modMessage = rcon.loadmodule('rconmessage', 'RconMessage')
+    modMessage.setInterval( config['repeatMessage']['interval'] )
+    modMessage.setMessages( config['repeatMessage']['messages'] )
+
+    if 'commands' in config:
+	modCommand = rcon.loadmodule('rconcommand', 'RconCommand')
+	_p = os.path.abspath(os.path.dirname(sys.argv[0]))
+        modCommand.setConfig( _p + '/' + config['commands'] )
+
+
+if GUI:
+    modGUI = rcon.loadmodule('rcongui', 'RconGUI')
 
 ##
 # Connect to server
@@ -83,4 +89,4 @@ modMessage.setMessages( config['repeatMessage']['messages'] )
 rcon.connect()
 
 # remove PID
-os.unlink(pidfile)
+if not(GUI): os.unlink(pidfile)
